tabs! and a *bit* of refactoring
