+++ fitch-proof/packages/app/package.json
@@ -11,6 +11,7 @@
   },
   "devDependencies": {
     "@stylistic/eslint-plugin": "^5.4.0",
+    "@types/wicg-file-system-access": "^2023.10.7",
     "@typescript-eslint/eslint-plugin": "^8.44.1",
     "@typescript-eslint/parser": "^8.44.0",
     "eslint": "^9.36.0",
@@ -24,8 +25,10 @@
     "@tailwindcss/vite": "^4.1.14",
     "@tsparticles/engine": "^3.9.1",
     "@tsparticles/preset-confetti": "^3.2.0",
+    "@types/alpinejs": "^3.13.11",
     "@workspace/library": "workspace:*",
     "@ym-han/monaco-error-lens": "^0.1.3",
+    "alpinejs": "^3.15.0",
     "tailwindcss": "^4.1.14",
     "vite-plugin-wasm": "^3.5.0"
   }
